requirements.txt
django
python-dotenv
djangorestframework
djangorestframework-simplejwt
celery
redis
requests
drf-yasg
django-cors-headers 